+++ Editor/Scripts/BasisImporter.cs
@@ -6,11 +6,7 @@ using System;
 using System.IO;
 using UnityEngine;
 using UnityEngine.Profiling;
-#if UNITY_2020_2_OR_NEWER
 using UnityEditor.AssetImporters;
-#else
-using UnityEditor.Experimental.AssetImporters;
-#endif
 
 namespace KtxUnity.Editor
 {
@@ -23,17 +19,15 @@ namespace KtxUnity.Editor
             var texture = new BasisUniversalTexture();
             var result = AsyncHelpers.RunSync(() =>
             {
-                using (var alloc = new ManagedNativeArray(File.ReadAllBytes(assetPath)))
-                {
-                    return texture.LoadFromBytesInternal(
-                        alloc.nativeArray,
-                        linear,
-                        layer,
-                        faceSlice,
-                        levelLowerLimit,
-                        importLevelChain
-                    );
-                }
+                using var alloc = new ManagedNativeArray(File.ReadAllBytes(assetPath));
+                return texture.LoadFromBytesInternal(
+                    alloc.nativeArray.AsReadOnly(),
+                    linear,
+                    layer,
+                    faceSlice,
+                    levelLowerLimit,
+                    importLevelChain
+                );
             });
             Profiler.EndSample();
             return result;