+++ Editor/Scripts/KtxImporterEditor.cs
@@ -3,11 +3,7 @@
 
 
 using UnityEditor;
-#if UNITY_2020_2_OR_NEWER
 using UnityEditor.AssetImporters;
-#else
-using UnityEditor.Experimental.AssetImporters;
-#endif
 
 namespace KtxUnity.Editor
 {